requirements.txt
selenium==4.18.1
behave==1.2.6
allure-behave==2.13.2
openpyxl==3.1.2
redis==5.0.1
webdriver-manager==4.0.1
Faker==22.2.0
sqlalchemy==2.0.25
Pillow==10.2.0
flask==3.0.2
flask-socketio==5.3.6
simple-websocket==1.0.0